+++ test/DotNetIsolated60/Program.cs
@@ -1,0 +1,11 @@
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Extensions.Hosting;
+
+//Debugger.Launch();
+
+var host = new HostBuilder()
+    .ConfigureFunctionsWorkerDefaults()
+    .ConfigureGeneratedFunctionMetadataProvider()
+    .Build();
+
+host.Run();