+++ release_notes.md
@@ -3,6 +3,9 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
-
-**Release sprint:** Sprint 142
-[ [bugs](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+143%22+label%3Abug+is%3Aclosed) | [features](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+143%22+label%3Afeature+is%3Aclosed) ]
+- Update Python Worker Version to [4.14.0](https://github.com/Azure/azure-functions-python-worker/releases/tag/4.14.0)
+- Update protobuf to [v1.9.0-protofile](https://github.com/Azure/azure-functions-language-worker-protobuf/releases/tag/v1.9.0-protofile)
+- Update PowerShell Worker 7.2 to 4.0.2803 [Release Note](https://github.com/Azure/azure-functions-powershell-worker/releases/tag/v4.0.2803)
+- Update PowerShell Worker 7.4 to 4.0.2802 [Release Note](https://github.com/Azure/azure-functions-powershell-worker/releases/tag/v4.0.2802)
+- Fixing bug with placeholder misses in dotnet-isolated #9253
+- Update PowerShell Worker 7.0 to 4.0.2823 [Release Note](https://github.com/Azure/azure-functions-powershell-worker/releases/tag/v4.0.2823)